moving config reading to js sdk
