+++ libs/js-sdk/src/utils/config/defaults.ts
@@ -1,0 +1,5 @@
+const configFile = "./pixeleye.config.js";
+const endpoint = "https://pixeleye.io";
+const port = "3003";
+
+export const defaults = { configFile, endpoint, port };
