+++ libs/js-sdk/src/utils/config/index.ts
@@ -1,0 +1,2 @@
+export * from "./defaults";
+export * from "./config";
